Build Skiko with Clang-CL (#1020)
Official Skia documentation suggests that it's [_highly
recommended_](https://skia.org/docs/user/build/#highly-recommended-build-with-clang-cl)
to build Skia with Clang-CL on Windows, and that this dramatically
improves Skia performance with Software rendering and in other areas.
This corresponds with my experience, so here it is.

Related Skia-pack PR: https://github.com/JetBrains/skia-pack/pull/64

Prerequisite: #1024

+++ skiko/src/awtMain/cpp/windows/InternalDirectXApi.cc
@@ -15,7 +15,6 @@
 #include "ganesh/GrBackendSurface.h"
 #include "ganesh/GrDirectContext.h"
 #include "SkSurface.h"
-#include "../common/interop.hh"
 
 #include "ganesh/d3d/GrD3DTypes.h"
 #include "ganesh/d3d/GrD3DBackendContext.h"
@@ -351,7 +350,7 @@ extern "C"
         dst.PlacedFootprint.Footprint.Depth = 1;
         dst.PlacedFootprint.Footprint.RowPitch = calculateRowPitch(texture->width);
 
-        D3D12_BOX srcBox = {0, 0, 0, texture->width, texture->height, 1};
+        D3D12_BOX srcBox = {0, 0, 0, static_cast<UINT>(texture->width), static_cast<UINT>(texture->height), 1};
 
         commandList->CopyTextureRegion(&dst, 0, 0, 0, &src, &srcBox);
 
@@ -381,7 +380,7 @@ extern "C"
         DirectXOffScreenTexture *texture = fromJavaPointer<DirectXOffScreenTexture *>(texturePtr);
 
         auto rangeLength = texture->readbackBufferWidth();
-        D3D12_RANGE readbackBufferRange{ 0, rangeLength };
+        D3D12_RANGE readbackBufferRange{ 0, static_cast<SIZE_T>(rangeLength) };
 
         /*
          * TODO: memcpy from unaligned texture is not supported, line by line copy is very slow,
